pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "humancrafter"
version = "0.1.0"
description = "Structured HumanCrafter research scaffold for inference and evaluation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "Pillow",
  "PyYAML",
]

[project.optional-dependencies]
torch = ["torch"]

[tool.setuptools]
include-package-data = true
